Added Kafka as output
Added a kafka module which will enable us to send output
to a kafka queue. Also modified the DAG that is created
for scheduling the execution. Had trouble running it under
PythonOperator, as loading of a Kafka module would fail.
So it is now executed as a BashOperator. Furthermore, some
changes in the logic of execution was made. before we were
converting it to dataframe and then checking if the value
is 100 or not. now we just check for next incremental value
to be available.